--- mission-lotto/src/main/java/com/lotto/web/exception/dto/ErrorResponse.java
@@ -1,4 +1,0 @@
-package com.lotto.web.exception.dto;
-
-public record ErrorResponse(String message) {
-}
