+++ mission-lotto/src/main/java/com/lotto/web/global/exception/exceptions/ErrorResponse.java
@@ -1,0 +1,4 @@
+package com.lotto.web.global.exception.exceptions;
+
+public record ErrorResponse(String message) {
+}
