+++ mission-lotto/src/main/java/com/lotto/web/lotto/domain/vo/exception/LottoRangeException.java
@@ -1,7 +1,7 @@
 package com.lotto.web.lotto.domain.vo.exception;
 
-import com.lotto.web.exception.domain.CustomErrorCode;
-import com.lotto.web.exception.domain.CustomException;
+import com.lotto.web.global.exception.exceptions.CustomErrorCode;
+import com.lotto.web.global.exception.exceptions.CustomException;
 
 public class LottoRangeException extends CustomException {
     public LottoRangeException() {
