+++ mission-lotto/src/main/java/com/lotto/web/lotto/service/exception/NotFoundLottoException.java
@@ -1,7 +1,7 @@
 package com.lotto.web.lotto.service.exception;
 
-import com.lotto.web.exception.domain.CustomErrorCode;
-import com.lotto.web.exception.domain.CustomException;
+import com.lotto.web.global.exception.exceptions.CustomErrorCode;
+import com.lotto.web.global.exception.exceptions.CustomException;
 
 public class NotFoundLottoException extends CustomException {
     public NotFoundLottoException() {
