+++ mission-lotto/src/main/java/com/lotto/web/member/service/exception/NotFoundMemberException.java
@@ -1,7 +1,7 @@
 package com.lotto.web.member.service.exception;
 
-import com.lotto.web.exception.domain.CustomErrorCode;
-import com.lotto.web.exception.domain.CustomException;
+import com.lotto.web.global.exception.exceptions.CustomErrorCode;
+import com.lotto.web.global.exception.exceptions.CustomException;
 
 public class NotFoundMemberException extends CustomException {
     public NotFoundMemberException() {
